creation and deletion of cities added

+++ src/app/shared/components/city-card/city-card.component.ts
@@ -1,8 +1,12 @@
 import { Component, Input, OnDestroy } from '@angular/core';
 import { City } from "../../../cities/models/city.model";
 import { CitiesDataService } from "../../../cities/services/cities-data.service";
-import { catchError, Subject, take, takeUntil, tap, throwError } from "rxjs";
+import { catchError, Subject, take, takeUntil, throwError } from "rxjs";
 import { CitiesStoreService } from "../../../cities/services/cities-store.service";
+import { MatDialog, MatDialogRef } from "@angular/material/dialog";
+import {
+  DeleteConfirmationComponent
+} from "../../../cities/components/delete-confirmation/delete-confirmation.component";
 
 @Component({
   selector: 'app-city-card',
@@ -17,6 +21,7 @@ export class CityCardComponent implements OnDestroy {
   constructor(
     private dataService: CitiesDataService,
     private dataStore: CitiesStoreService,
+    public dialog: MatDialog,
   ) {
   }
 
@@ -41,12 +46,18 @@ export class CityCardComponent implements OnDestroy {
   }
 
   deleteCity() {
-    this.dataService.deleteCity(this.city.id)
-      .pipe(
-        take(1),
-        takeUntil(this.notifier$),
-        catchError((err) => throwError(() => err))
-      )
-      .subscribe(() => this.dataStore.deleteCity(this.city.id));
+    const dialogRef: MatDialogRef<DeleteConfirmationComponent> = this.dialog.open(DeleteConfirmationComponent,
+      {width: '250px'});
+    dialogRef.afterClosed().subscribe(res => {
+      if (!!res) {
+        this.dataService.deleteCity(this.city.id)
+          .pipe(
+            take(1),
+            takeUntil(this.notifier$),
+            catchError((err) => throwError(() => err))
+          )
+          .subscribe(() => this.dataStore.deleteCity(this.city.id));
+      }
+    });
   }
 }
